fix(clippy): auto-fix batch for tests/benches/examples (60 files)
cargo clippy --fix across all-targets. 2 pre-existing broken tests
(bench_102_tilert_pipeline_goat, depth_invariance_micro_belief) block
--fix for the remaining test targets — they reference removed modules
(weights, audit_depth_invariance) from sibling-agent refactoring and
are unrelated to clippy.

+++ tests/bench_299_engram_goat.rs
@@ -364,7 +364,7 @@ fn spearman_rho(a: &[(usize, f32)], b: &[(usize, f32)]) -> f64 {
     let mut n = 0u64;
     for (id, ra) in &rank_a {
         if let Some(rb) = rank_b.get(id) {
-            let (x, y) = (*ra as f64, *rb as f64);
+            let (x, y) = ((*ra), (*rb));
             sum_xy += x * y;
             sum_x += x;
             sum_y += y;
